+++ src/components/organisms/ErrorBoundary/components/ErrorFallback/index.test.tsx
@@ -16,9 +16,14 @@ describe('ErrorFallback', () => {
 		expect(getByText('Show error details')).toBeTruthy();
 	});
 
-	it('renders error message if provided', () => {
-		const {getByText} = render(<ErrorFallback error="This is an error" />);
+	it('renders error message if provided and user is in debug mode', () => {
+		const {getByText} = render(<ErrorFallback error="This is an error" isDebug />);
 
 		expect(getByText('This is an error')).toBeTruthy();
+	});
+
+	it('does NOT render error message when NOT in debug mode', () => {
+		const {queryByText} = render(<ErrorFallback error="This is an error" />);
+		expect(queryByText('This is an error')).toBeNull();
 	});
 });