+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- an error with list component in which props were not passing properly - [JUIP-171](https://janiscommerce.atlassian.net/browse/JUIP-171)
+
 ## [1.21.0] - 2025-07-07
 
 ### Added